Ad hoc payment: Single fee
This is a catalogue-style version of ad hoc payment, with only one item. There is a set fee to pay. In start-2, user can decide to pay for more then one item.  

User receives a GOV.UK Pay reference number to include with their offline application as ‘proof of payment’.